funziona prima parte ricalcoli

+++ src/main/java/com/deltasi/elezioni/contracts/ISezioneService.java
@@ -10,5 +10,5 @@ public interface ISezioneService {
     Sezione findByNumerosezioneAndTipoelezioneId(Integer numeroSezione, Integer tipoElezioneId);
     List<Sezione> findAllBy();
     List<Integer> countAllByTipoelezioneIdAndTipoelezioneIdIn(int tipoElezione, int tipoElezione1);
-    List<Integer> countAllByTipoelezioneIdAndMunicipioAndTipoelezioneIdIn(int tipoElezioneId, int municipio,int tipoElezione1);
+    List<Long> countAllByTipoelezioneIdAndMunicipioAndTipoelezioneIdIn(int tipoElezioneId, int municipio,int tipoElezione1);
 }
