+++ src/main/java/com/deltasi/elezioni/repository/SezioneDAO.java
@@ -13,7 +13,7 @@ public interface SezioneDAO extends JpaRepository<Sezione, Long> {
     Sezione findByNumerosezioneAndTipoelezioneId(Integer numeroSezione, Integer tipoElezioneId);
     List<Sezione> findAllBy();
     List<Integer> countAllByTipoelezioneIdAndTipoelezioneIdIn(int tipoElezione, int tipoElezione1);
-    List<Integer> countAllByTipoelezioneIdAndMunicipioAndTipoelezioneIdIn(int tipoElezioneId, int municipio,int tipoElezione1);
+    List<Long> countAllByTipoelezioneIdAndMunicipioAndTipoelezioneIdIn(int tipoElezioneId, int municipio,int tipoElezione1);
 
 
 }
